Split indicators page into page and targets tab
The indicators page object was getting large and unwieldy.
I've refactored all of the functionality specific to the
targets tab into its own object. The result is an additional
object model for the targets tab. The targets tab object uses
a few of the methods from the indicators page object but is
otherwise surprisingly independent of its parent page. The
result is objects that are easier to understand and tests
that don't have WebDriver-specific code in them. Hopefully,
this will make long-term maintenance easier and the tests
more robust.
