+++ test/js/test/specs/target.js
@@ -2,6 +2,7 @@ var assert = require('chai').assert;
 var expect = require('chai').expect;
 var LoginPage = require('../pages/login.page.js');
 var IndPage = require('../pages/indicators.page.js');
+var TargetsTab = require('../pages/targets.page.js');
 var util = require('../lib/testutil.js');
 
 describe('Indicator Targets', function() {
@@ -21,39 +22,39 @@ describe('Indicator Targets', function() {
   describe('Create new indicator dialog', function() {
     this.timeout(0);
     it('should open new indicator form when button is clicked', function() {
-      IndPage.clickNewIndicatorButton();
-      assert.equal('Create an Indicator', IndPage.pageName());
+      TargetsTab.clickNewIndicatorButton();
+      assert.equal('Create an Indicator', TargetsTab.pageName());
     });
 
     it('should save the new indicator when the button is clicked', function() {
-      IndPage.saveNewIndicator();
+      TargetsTab.saveNewIndicator();
     });
 
     it('should have a "Reset" button', function() {
-      IndPage.clickResetButton();
+      TargetsTab.clickResetButton();
     });
 
     it('should clear form when "Reset" button clicked', function() {
-      IndPage.setIndicatorName('Testing the reset button');
-      IndPage.setUnitOfMeasure('Clicks');
-      IndPage.setLoPTarget(38);
-      IndPage.setBaseline(39);
-      IndPage.setTargetFrequency('Life of Program (LoP) only');
-      IndPage.clickResetButton();
-      assert.equal('Temporary', IndPage.getIndicatorName());
-      assert.equal('', IndPage.getUnitOfMeasure());
-      assert.equal('', IndPage.getLoPTarget());
-      assert.equal('', IndPage.getBaseline());
-      assert.equal('', IndPage.getTargetFrequency());
+      TargetsTab.setIndicatorName('Testing the reset button');
+      TargetsTab.setUnitOfMeasure('Clicks');
+      TargetsTab.setLoPTarget(38);
+      TargetsTab.setBaseline(39);
+      TargetsTab.setTargetFrequency('Life of Program (LoP) only');
+      TargetsTab.clickResetButton();
+      assert.equal('Temporary', TargetsTab.getIndicatorName());
+      assert.equal('', TargetsTab.getUnitOfMeasure());
+      assert.equal('', TargetsTab.getLoPTarget());
+      assert.equal('', TargetsTab.getBaseline());
+      assert.equal('', TargetsTab.getTargetFrequency());
     });
 
     it('should save data when "Save changes" button clicked', function() {
-      IndPage.setIndicatorName('Testing the save changes button');
-      IndPage.setUnitOfMeasure('Dashes');
-      IndPage.setLoPTarget(52);
-      IndPage.setBaseline(53);
-      IndPage.setTargetFrequency('Life of Program (LoP) only');
-      IndPage.saveIndicatorChanges();
+      TargetsTab.setIndicatorName('Testing the save changes button');
+      TargetsTab.setUnitOfMeasure('Dashes');
+      TargetsTab.setLoPTarget(52);
+      TargetsTab.setBaseline(53);
+      TargetsTab.setTargetFrequency('Life of Program (LoP) only');
+      TargetsTab.saveIndicatorChanges();
     });
   }); // end new indicator dialog
 
@@ -63,65 +64,65 @@ describe('Indicator Targets', function() {
     it('should require "Unit of measure"', function() {
       // Set everything but the unit of measure then try to save
       IndPage.clickIndicatorsLink();
-      IndPage.clickNewIndicatorButton();
-      IndPage.saveNewIndicator();
-      IndPage.setIndicatorName('Unit of measure test');
-      IndPage.setLoPTarget(68);
-      IndPage.setBaseline(69);
-      IndPage.setTargetFrequency('Life of Program (LoP) only');
-      IndPage.saveIndicatorChanges();
+      TargetsTab.clickNewIndicatorButton();
+      TargetsTab.saveNewIndicator();
+      TargetsTab.setIndicatorName('Unit of measure test');
+      TargetsTab.setLoPTarget(68);
+      TargetsTab.setBaseline(69);
+      TargetsTab.setTargetFrequency('Life of Program (LoP) only');
+      TargetsTab.saveIndicatorChanges();
       // Should get this error message
       assert.equal('Please complete all required fields in the Targets tab.',
-                   IndPage.getAlertMsg(),
+                   TargetsTab.getAlertMsg(),
                    'Did not receive expected failure message');
     });
 
     it('should require "Life of Program (LoP) target" value', function() {
       // Set everything but the LoP target then try to save
       IndPage.clickIndicatorsLink();
-      IndPage.clickNewIndicatorButton();
-      IndPage.saveNewIndicator();
-      IndPage.setUnitOfMeasure('LoP target test');
-      IndPage.setBaseline(84);
-      IndPage.setTargetFrequency('Life of Program (LoP) only');
-      IndPage.saveIndicatorChanges();
+      TargetsTab.clickNewIndicatorButton();
+      TargetsTab.saveNewIndicator();
+      TargetsTab.setUnitOfMeasure('LoP target test');
+      TargetsTab.setBaseline(84);
+      TargetsTab.setTargetFrequency('Life of Program (LoP) only');
+      TargetsTab.saveIndicatorChanges();
       // Should get this error message
       assert.equal('Please complete all required fields in the Targets tab.',
-                   IndPage.getAlertMsg(),
+                   TargetsTab.getAlertMsg(),
                    'Did not receive expected failure message');
     });
 
     it('should require target Baseline if "Not applicable" not checked', function() {
       // Set everything but the baseline value then try to save
       IndPage.clickIndicatorsLink();
-      IndPage.clickNewIndicatorButton();
-      IndPage.saveNewIndicator();
-      IndPage.setUnitOfMeasure('Baseline value test');
-      IndPage.setLoPTarget(99);
-      IndPage.setTargetFrequency('Life of Program (LoP) only');
-      IndPage.saveIndicatorChanges();
+      TargetsTab.clickNewIndicatorButton();
+      TargetsTab.saveNewIndicator();
+      TargetsTab.setUnitOfMeasure('Baseline value test');
+      TargetsTab.setLoPTarget(99);
+      TargetsTab.setTargetFrequency('Life of Program (LoP) only');
+      TargetsTab.saveIndicatorChanges();
       // Should get this error message
       assert.equal('Please complete all required fields in the Targets tab.',
-                   IndPage.getAlertMsg(),
+                   TargetsTab.getAlertMsg(),
                    'Did not receive expected failure message');
     });
 
     it('should *not* require target Baseline if "Not applicable" *is* checked', function() {
       // Set everything but the baseline value
       IndPage.clickIndicatorsLink();
-      IndPage.clickNewIndicatorButton();
-      IndPage.saveNewIndicator();
-      IndPage.setUnitOfMeasure('Baseline value not applicable test');
-      IndPage.setLoPTarget(114);
-      IndPage.setTargetFrequency('Life of Program (LoP) only');
+      TargetsTab.clickNewIndicatorButton();
+      TargetsTab.saveNewIndicator();
+      TargetsTab.setUnitOfMeasure('Baseline value not applicable test');
+      TargetsTab.setLoPTarget(114);
+      TargetsTab.setTargetFrequency('Life of Program (LoP) only');
       // check "Not applicable" then try to save
-      IndPage.setBaselineNA();
-      IndPage.saveIndicatorChanges();
+      TargetsTab.setBaselineNA();
+      TargetsTab.saveIndicatorChanges();
       // Should get this success message
       // FIXME: This test fails because the success message is too fast;
       //        I don't know how to catch it with WebDriver
       assert.equal('Success, form data saved.',
-                   IndPage.getAlertMsg(),
+                   TargetsTab.getAlertMsg(),
                    'Did not receive expected success message');
       // Should not get this message
       // FIXME: I'd rather not use the negative logic; much prefer the
@@ -134,15 +135,15 @@ describe('Indicator Targets', function() {
     it('should require Target frequency', function() {
       // Set everything but the target frequency then try to save
       IndPage.clickIndicatorsLink();
-      IndPage.clickNewIndicatorButton();
-      IndPage.saveNewIndicator();
-      IndPage.setUnitOfMeasure('Target frequency test');
-      IndPage.setLoPTarget(139);
-      IndPage.setBaseline(140);
-      IndPage.saveIndicatorChanges();
+      TargetsTab.clickNewIndicatorButton();
+      TargetsTab.saveNewIndicator();
+      TargetsTab.setUnitOfMeasure('Target frequency test');
+      TargetsTab.setLoPTarget(139);
+      TargetsTab.setBaseline(140);
+      TargetsTab.saveIndicatorChanges();
       // Should get this error message
       assert.equal('Please complete all required fields in the Targets tab.',
-                   IndPage.getAlertMsg(),
+                   TargetsTab.getAlertMsg(),
                    'Did not receive expected failure message');
     });
 
@@ -151,30 +152,30 @@ describe('Indicator Targets', function() {
     describe('"Life of Program (LoP) only" target frequency', function() {
       it('should permit only numeric values for LoP target', function() {
         IndPage.clickIndicatorsLink();
-        IndPage.clickNewIndicatorButton();
-        IndPage.saveNewIndicator();
+        TargetsTab.clickNewIndicatorButton();
+        TargetsTab.saveNewIndicator();
 
-        IndPage.setIndicatorName('LoP only target testing');
-        IndPage.setUnitOfMeasure('Furlongs per fortnight');
-        IndPage.setLoPTarget(158);
-        IndPage.setBaseline(156);
-        IndPage.setTargetFrequency('Life of Program (LoP) only');
+        TargetsTab.setIndicatorName('LoP only target testing');
+        TargetsTab.setUnitOfMeasure('Furlongs per fortnight');
+        TargetsTab.setLoPTarget(158);
+        TargetsTab.setBaseline(156);
+        TargetsTab.setTargetFrequency('Life of Program (LoP) only');
         // This should succeed
-        IndPage.saveIndicatorChanges();
+        TargetsTab.saveIndicatorChanges();
       });
 
       it('should reject non-numeric values for LoP target', function() {
-        IndPage.setLoPTarget('"166"');
+        TargetsTab.setLoPTarget('"166"');
         // This should fail
-        IndPage.saveIndicatorChanges();
-        let errorHint = IndPage.getLoPErrorHint();
+        TargetsTab.saveIndicatorChanges();
+        let errorHint = TargetsTab.getLoPErrorHint();
         assert(errorHint.includes('Please enter a number larger than zero'),
                'Did not receive expected failure message');
         // Make it numeric; this should succeed
-        IndPage.setLoPTarget(173);
-        IndPage.saveIndicatorChanges();
+        TargetsTab.setLoPTarget(173);
+        TargetsTab.saveIndicatorChanges();
         assert.equal(173,
-                     IndPage.getLoPTarget(),
+                     TargetsTab.getLoPTarget(),
                      'Did not receive expected value from getLoPTarget()');
       });
