+++ test/js/wdio.conf.js
@@ -43,7 +43,7 @@ exports.config = {
         // grid with only 5 firefox instances available you can make sure that not more than
         // 5 instances get started at a time.
         maxInstances: 5,
-        browserName: 'firefox'
+        browserName: 'chrome'
     }],
     //
     // ===================
